wip: buffer replication.
Thinking about changing the name from "controller" to "manager".
Shorter, goes better with the "worker", and since there is also the
"client", it makes for better single letter abbreviations (m, w, c).

Decided that on a node, a tenant will have at most 1 of each: manager,
worker, client. This makes for much simpler routing with
"/tenants/-/manager" etc, as opposed to "/tenants/-/nodes/{ID}". Makes
it much less of a pain to curl, because the {ID} changes on every node
restart, but the "new" routes stay the same.

Data directory now split into "manager" and "worker" subdirectories, to
make it easier to see which data is where. This means that the directory
layout matches the routes: "tenenants/-/manager/topics" is the same
route and path on disk; same for "tenants/-/worker/buffers". Ha!