wip: buffer replication.
Buffers now replicated asynchronously. The idea is to avoid two modes:
sync and async. If there is a need for "sync" replication, the buffer
can just wait for async replicas to catch up before returning ok (this
is not yet implemented). TODO: clean shutdown: make sure all primaries
shut down before replicas; maybe not a big deal for now?

On startup, replicas resume from the point where they left. TODO: when
starting a new replica for an existing buffer, allow it to start from
the beginning, of from tail: there may be no need to replicate from the
very beginning. See NOTES; the basic idea is that when the cluster is in
trouble and nodes are failing, the last thing it needs is replicating
everything from the start.