wip: buffer replication.
Each message in a segment not has a sha256 which is the previous sha +
the sha of the serialized message. This can be used to check the
segment's integrity (and is done by segment.verify() method, which at
this point is pretty inefficient). So this brings up the question of
what to do with failed segments / buffers; will need to update
controller code to handle "health" proactively; this also means that the
controller will need to persist replica information (right now it is the
buffer that maintains the info on its replicas; but what if that buffer
failes to come up? the controller should have the authoritative list of
replicas).

This code is pretty janky; but get it working first, then make it nice.

Also: buffer writes now synchronous again: no need for channels.
Simplified segment code.