wip: buffer replication.
Better division of responsibilities between buffers and segments / less
duplication. Buffers are now responsible for the monotonic message IDs,
and for checksumming. Segments don't know the context in which they
operate: first message in each segment is accessed at index 0. Before,
it was the segments which were translating from the message id to the
offset in the segment; this meant that info had to be passed back to the
buffer, and this made things messy.