wip: buffer replication.
The manager now keeps track of replicas (used to be the buffer). This
means that when the buffer files are lost, replication information is
not (because the manager has it). Simplified buffer registration (it is
now separate from replica assignment). Still very simplistic, but trying
to not trip too much about the whatifs right now.