wip: buffer replication.
Added ability to start a new buffer at start offset other than 0. This
is for situations where a replica is created and it doesn't start at 0
either because earlier segments were trimmed, or the old segments are
not replicated for performance reasons.

Also, consumer will now "skip" over gaps in message ids. This is to deal
with situations where a segment has been trimmed, and the consumer
became stale. It will now skip ahead to the next available message.